fix: split code based on responsibility
