+++ gql/example/simple/schema/hello.graphql
@@ -1,0 +1,3 @@
+{
+    name
+}
