+++ gql/http_client.go
@@ -6,6 +6,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
+	"log"
 	"mime/multipart"
 	"net/http"
 
@@ -25,10 +26,12 @@ type Client struct {
 	// To log to standard out, use:
 	//  client.Log = func(s string) { log.Println(s) }
 	Log func(s string)
+
+	Debug bool
 }
 
 // NewClient makes a new Client capable of making GraphQL requests.
-func NewClient(endpoint string, opts ...ClientOption) *Client {
+func NewHttpClient(endpoint string, opts ...ClientOption) *Client {
 	c := &Client{
 		endpoint: endpoint,
 		Log:      func(string) {},
@@ -51,7 +54,7 @@ func (c *Client) logf(format string, args ...interface{}) {
 // Pass in a nil response object to skip response parsing.
 // If the request fails or the server returns an error, the first error
 // will be returned.
-func (c *Client) Run(ctx context.Context, req *Request, resp interface{}) error {
+func (c *Client) Run(ctx context.Context, req *Request, resp any) error {
 	select {
 	case <-ctx.Done():
 		return ctx.Err()
@@ -66,7 +69,7 @@ func (c *Client) Run(ctx context.Context, req *Request, resp interface{}) error 
 	return c.runWithJSON(ctx, req, resp)
 }
 
-func (c *Client) runWithJSON(ctx context.Context, req *Request, resp interface{}) error {
+func (c *Client) runWithJSON(ctx context.Context, req *Request, resp any) error {
 	var requestBody bytes.Buffer
 	reqBodyObject := struct {
 		Query     string                 `json:"query"`
@@ -78,8 +81,15 @@ func (c *Client) runWithJSON(ctx context.Context, req *Request, resp interface{}
 	if err := json.NewEncoder(&requestBody).Encode(reqBodyObject); err != nil {
 		return errors.Wrap(err, "encode body")
 	}
-	c.logf(">> variables: %v", req.vars)
-	c.logf(">> query: %s", req.q)
+
+	log.Println("resprespresp")
+	log.Println(resp)
+	log.Println("resprespresp")
+
+	if c.Debug {
+		c.logf(">> variables: %v", req.vars)
+		c.logf(">> query: %s", req.q)
+	}
 
 	r, err := http.NewRequest(http.MethodPost, c.endpoint, &requestBody)
 	if err != nil {
@@ -111,7 +121,7 @@ func (c *Client) runWithJSON(ctx context.Context, req *Request, resp interface{}
 	return nil
 }
 
-func (c *Client) runWithPostFields(ctx context.Context, req *Request, resp interface{}) error {
+func (c *Client) runWithPostFields(ctx context.Context, req *Request, resp any) error {
 	var requestBody bytes.Buffer
 	writer := multipart.NewWriter(&requestBody)
 	if err := writer.WriteField("query", req.q); err != nil {
@@ -139,9 +149,13 @@ func (c *Client) runWithPostFields(ctx context.Context, req *Request, resp inter
 	if err := writer.Close(); err != nil {
 		return errors.Wrap(err, "close writer")
 	}
-	c.logf(">> variables: %s", variablesBuf.String())
-	c.logf(">> files: %d", len(req.files))
-	c.logf(">> query: %s", req.q)
+
+	if c.Debug {
+		c.logf(">> variables: %s", variablesBuf.String())
+		c.logf(">> files: %d", len(req.files))
+		c.logf(">> query: %s", req.q)
+	}
+
 	gr := &graphResponse{
 		Data: resp,
 	}
@@ -157,7 +171,11 @@ func (c *Client) runWithPostFields(ctx context.Context, req *Request, resp inter
 			r.Header.Add(key, value)
 		}
 	}
-	c.logf(">> headers: %v", r.Header)
+
+	if c.Debug {
+		c.logf(">> headers: %v", r.Header)
+	}
+
 	r = r.WithContext(ctx)
 	res, err := c.httpClient.Do(r)
 	if err != nil {
@@ -168,7 +186,11 @@ func (c *Client) runWithPostFields(ctx context.Context, req *Request, resp inter
 	if _, err := io.Copy(&buf, res.Body); err != nil {
 		return errors.Wrap(err, "reading body")
 	}
-	c.logf("<< %s", buf.String())
+
+	if c.Debug {
+		c.logf("<< %s", buf.String())
+	}
+
 	if err := json.NewDecoder(&buf).Decode(&gr); err != nil {
 		if res.StatusCode != http.StatusOK {
 			return fmt.Errorf("graphql: server returned a non-200 status code: %v", res.StatusCode)
@@ -180,107 +202,4 @@ func (c *Client) runWithPostFields(ctx context.Context, req *Request, resp inter
 		return gr.Errors[0]
 	}
 	return nil
-}
-
-// WithHTTPClient specifies the underlying http.Client to use when
-// making requests.
-//
-//	NewClient(endpoint, WithHTTPClient(specificHTTPClient))
-func WithHTTPClient(httpclient *http.Client) ClientOption {
-	return func(client *Client) {
-		client.httpClient = httpclient
-	}
-}
-
-// UseMultipartForm uses multipart/form-data and activates support for
-// files.
-func UseMultipartForm() ClientOption {
-	return func(client *Client) {
-		client.useMultipartForm = true
-	}
-}
-
-// ImmediatelyCloseReqBody will close the req body immediately after each request body is ready
-func ImmediatelyCloseReqBody() ClientOption {
-	return func(client *Client) {
-		client.closeReq = true
-	}
-}
-
-// ClientOption are functions that are passed into NewClient to
-// modify the behaviour of the Client.
-type ClientOption func(*Client)
-
-type graphErr struct {
-	Message string
-}
-
-func (e graphErr) Error() string {
-	return "graphql: " + e.Message
-}
-
-type graphResponse struct {
-	Data   interface{}
-	Errors []graphErr
-}
-
-// Request is a GraphQL request.
-type Request struct {
-	q     string
-	vars  map[string]interface{}
-	files []File
-
-	// Header represent any request headers that will be set
-	// when the request is made.
-	Header http.Header
-}
-
-// NewRequest makes a new Request with the specified string.
-func NewRequest(q string) *Request {
-	req := &Request{
-		q:      q,
-		Header: make(map[string][]string),
-	}
-	return req
-}
-
-// Var sets a variable.
-func (req *Request) Var(key string, value interface{}) {
-	if req.vars == nil {
-		req.vars = make(map[string]interface{})
-	}
-	req.vars[key] = value
-}
-
-// Vars gets the variables for this Request.
-func (req *Request) Vars() map[string]interface{} {
-	return req.vars
-}
-
-// Files gets the files in this request.
-func (req *Request) Files() []File {
-	return req.files
-}
-
-// Query gets the query string of this request.
-func (req *Request) Query() string {
-	return req.q
-}
-
-// File sets a file to upload.
-// Files are only supported with a Client that was created with
-// the UseMultipartForm option.
-func (req *Request) File(fieldname, filename string, r io.Reader) {
-	req.files = append(req.files, File{
-		Field: fieldname,
-		Name:  filename,
-		R:     r,
-	})
-}
-
-// File represents a file to upload.
-type File struct {
-	Field string
-	Name  string
-	R     io.Reader
 }
