+++ gql/http_options.go
@@ -1,0 +1,39 @@
+package gql
+
+import "net/http"
+
+// ClientOption are functions that are passed into NewClient to
+// modify the behaviour of the Client.
+type ClientOption func(*Client)
+
+// WithHTTPClient specifies the underlying http.Client to use when
+// making requests.
+//
+//	NewClient(endpoint, WithHTTPClient(specificHTTPClient))
+func WithHTTPClient(httpclient *http.Client) ClientOption {
+	return func(client *Client) {
+		client.httpClient = httpclient
+	}
+}
+
+// UseMultipartForm uses multipart/form-data and activates support for
+// files.
+func UseMultipartForm() ClientOption {
+	return func(client *Client) {
+		client.useMultipartForm = true
+	}
+}
+
+// ImmediatelyCloseReqBody will close the req body immediately after each request body is ready
+func ImmediatelyCloseReqBody() ClientOption {
+	return func(client *Client) {
+		client.closeReq = true
+	}
+}
+
+// Debug will print graphql call information
+func Debug() ClientOption {
+	return func(client *Client) {
+		client.Debug = true
+	}
+}
