+++ gql/request.go
@@ -1,0 +1,67 @@
+package gql
+
+import (
+	"io"
+	"net/http"
+)
+
+// Request is a GraphQL request.
+type Request struct {
+	q     string
+	vars  map[string]interface{}
+	files []File
+
+	// Header represent any request headers that will be set
+	// when the request is made.
+	Header http.Header
+}
+
+// File represents a file to upload.
+type File struct {
+	Field string
+	Name  string
+	R     io.Reader
+}
+
+// NewRequest makes a new Request with the gql query string.
+func NewRequest(q string) *Request {
+	req := &Request{
+		q:      q,
+		Header: make(map[string][]string),
+	}
+	return req
+}
+
+// Var sets a variable.
+func (req *Request) Var(key string, value interface{}) {
+	if req.vars == nil {
+		req.vars = make(map[string]interface{})
+	}
+	req.vars[key] = value
+}
+
+// Vars gets the variables for this Request.
+func (req *Request) Vars() map[string]interface{} {
+	return req.vars
+}
+
+// Files gets the files in this request.
+func (req *Request) Files() []File {
+	return req.files
+}
+
+// Query gets the query string of this request.
+func (req *Request) Query() string {
+	return req.q
+}
+
+// File sets a file to upload.
+// Files are only supported with a Client that was created with
+// the UseMultipartForm option.
+func (req *Request) File(fieldname, filename string, r io.Reader) {
+	req.files = append(req.files, File{
+		Field: fieldname,
+		Name:  filename,
+		R:     r,
+	})
+}
